+++ lib/server/http_server.dart
@@ -53,6 +53,10 @@ Future<void> runHttpServer(SendPort sendPort) async {
 
       // Add the vehicle to the ExVe model
       exveModel.addVehicle(vehicleSimulator.info.autoID);
+
+      // And a fleet
+      final fleet = Fleet.generate(vehicles: [vehicleSimulator.info.autoID]);
+      exveModel.addFleet(fleet);
     });
 
     final vehicleSimulatorApi = VehicleSimulatorApi(